+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Update reference from arXiv preprint to PRAB publication (see #166) (@jank324)
 - Rename converter modules to the respective name of the accelerator code (see #167) (@jank324)
 - Added imports to the code example in the README (see #188) (@jank324)
+- Refactor definitions of physical constants (see #189) (@hespe)
 
 ## [v0.6.3](https://github.com/desy-ml/cheetah/releases/tag/v0.6.3) (2024-03-28)